チーム内レビュー修正
Integer→Long

+++ backend/src/main/java/in/tech_camp/prototype_d/repository/AffiliationRepository.java
@@ -10,10 +10,10 @@ import in.tech_camp.prototype_d.entity.AffiliationEntity;
 @Mapper
 public interface AffiliationRepository {
     @Select("SELECT id FROM affiliations WHERE affiliation = #{affiliationName}")
-    Integer findIdByName(String affiliationName);
+    Long findIdByName(String affiliationName);
 
     @Select("SELECT affiliation FROM affiliations WHERE id = #{id}")
-    String findNameById(Integer id);
+    String findNameById(Long id);
 
     @Insert("INSERT INTO affiliations (affiliation) VALUES (#{affiliationName})")
     @Options(useGeneratedKeys = true, keyProperty = "id")
