+++ backend/src/main/java/in/tech_camp/prototype_d/service/UserService.java
@@ -35,7 +35,7 @@ public class UserService {
       public UserEntity registerUser(UserForm userForm, HttpServletRequest request, HttpServletResponse response) {
 
           //所属（affiliation）テーブルへ登録
-          Integer affiliationId = affiliationRepository.findIdByName(userForm.getAffiliation());
+          Long affiliationId = affiliationRepository.findIdByName(userForm.getAffiliation());
           if (affiliationId == null) {
               AffiliationEntity affiliation = new AffiliationEntity();
               affiliation.setAffiliationName(userForm.getAffiliation());
@@ -44,7 +44,7 @@ public class UserService {
           }
           
           //役職（Position）テーブルへ登録
-          Integer positionId = positionRepository.findIdByName(userForm.getPosition());
+          Long positionId = positionRepository.findIdByName(userForm.getPosition());
           if (positionId == null) {
               PositionEntity position = new PositionEntity();
               position.setPositionName(userForm.getPosition());
@@ -78,7 +78,7 @@ public class UserService {
 
 
   // Dtoに詰め替える
-    public UserDetailDto getUser(Integer userId) {
+    public UserDetailDto getUser(Long userId) {
 
     UserEntity userEntity = userRepository.findById(userId);
       UserDetailDto dto = new UserDetailDto();
